Lambda timer now stay on the main thread.
When changing screens if cleanup is true then all lambda timers are deleted. This is because it is not possible to check if the calling object still exists (weirdly the code will still run without the calling object but I'm worried about potential crashes). I could change this so they require a screen or actor and thus it would know if the calling object is still around but that feels a bit jank.